@@ -1,17 +1,26 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <LangVersion>9.0</LangVersion>
-	  <TargetFramework>net5.0</TargetFramework>
+	  <!--<LangVersion>9.0</LangVersion>
+	  <TargetFramework>net5.0</TargetFramework>-->
+	  <TargetFramework>net8.0</TargetFramework>
+	  <ImplicitUsings>enable</ImplicitUsings>
+	  <Nullable>disable</Nullable>
+
   </PropertyGroup>
-	<ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.1" />
+    <PackageReference Include="ConfigurationLibrary" Version="1.0.1" />
+  </ItemGroup>
+	<!--<ItemGroup>
 		<PackageReference Include="ConfigurationLibrary" Version="1.0.1" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="5.0.17" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.9" />
-	</ItemGroup>
+	</ItemGroup>-->
 
 	<ItemGroup>
 	  <ProjectReference Include="..\EntityCoreExtensions\EntityCoreExtensions.csproj" />
