@@ -1,8 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<LangVersion>9.0</LangVersion>
-		<TargetFramework>net5.0</TargetFramework>
+		<!--<LangVersion>9.0</LangVersion>
+		<TargetFramework>net5.0</TargetFramework>-->
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>disable</Nullable>
 		<IsPackable>false</IsPackable>
 	</PropertyGroup>
 
