@@ -1,51 +1,48 @@
 ﻿// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
-using CustomerDatabaseLibraryEntityFramework.Data.Configurations;
+
 using CustomerDatabaseLibraryEntityFramework.Models;
-using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore;
-using System;
 using ConfigurationLibrary.Classes;
 
 #nullable disable
 
 #nullable disable
 
-namespace CustomerDatabaseLibraryEntityFramework.Data
+namespace CustomerDatabaseLibraryEntityFramework.Data;
+
+public partial class CustomerContext : DbContext
 {
-    public partial class CustomerContext : DbContext
+    public CustomerContext()
     {
-        public CustomerContext()
-        {
-        }
-
-        public CustomerContext(DbContextOptions<CustomerContext> options)
-            : base(options)
-        {
-        }
-
-        public virtual DbSet<ContactTypes> ContactTypes { get; set; }
-        public virtual DbSet<Customer> Customer { get; set; }
-        public virtual DbSet<Genders> Genders { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-                // added by Karen
-                optionsBuilder.UseSqlServer(ConfigurationHelper.ConnectionString())
-                    .UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking);
-            }
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.HasAnnotation("Relational:Collation", "SQL_Latin1_General_CP1_CI_AS");
-
-            modelBuilder.ApplyConfiguration(new Configurations.ContactTypesConfiguration());
-            modelBuilder.ApplyConfiguration(new Configurations.CustomerConfiguration());
-            OnModelCreatingPartial(modelBuilder);
-        }
-
-        partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
     }
+
+    public CustomerContext(DbContextOptions<CustomerContext> options)
+        : base(options)
+    {
+    }
+
+    public virtual DbSet<ContactTypes> ContactTypes { get; set; }
+    public virtual DbSet<Customer> Customer { get; set; }
+    public virtual DbSet<Genders> Genders { get; set; }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        if (!optionsBuilder.IsConfigured)
+        {
+            // added by Karen
+            optionsBuilder.UseSqlServer(ConfigurationHelper.ConnectionString())
+                .UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking);
+        }
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.HasAnnotation("Relational:Collation", "SQL_Latin1_General_CP1_CI_AS");
+
+        modelBuilder.ApplyConfiguration(new Configurations.ContactTypesConfiguration());
+        modelBuilder.ApplyConfiguration(new Configurations.CustomerConfiguration());
+        OnModelCreatingPartial(modelBuilder);
+    }
+
+    partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
 }
